Oliver final

+++ src/com/edu/neu/csye7374/finalProject/models/InsuraceStrategyAPI.java
@@ -1,5 +1,5 @@
 package com.edu.neu.csye7374.finalProject.models;
 
 public interface InsuraceStrategyAPI {
-	public float calcuateBill(float preInsuranceBill);
+	public double calcuateBill(double preInsuranceBill);
 }
